tests/libraries/data/test_abstract_file_handler.py
import unittest
from data.source.file.handlers.abstract_file_handler import AbstractFileHandler
from unittest.mock import MagicMock

class ConcreteFileHandler(AbstractFileHandler):
    def read(self):
        return "test data"

    def save(self, data):
        pass

class TestAbstractFileHandler(unittest.TestCase):

    def test_read_method(self):
        """Test the read method of a concrete implementation."""
        handler = ConcreteFileHandler("dummy.txt")
        self.assertEqual(handler.read(), "test data")

    def test_save_method(self):
        """Test the save method of a concrete implementation."""
        handler = ConcreteFileHandler("dummy.txt")
        mock_data = MagicMock()
        handler.save(mock_data)
        mock_data.assert_not_called()  # As save is a pass in ConcreteFileHandler

    def test_is_abstract_class(self):
        """Test that AbstractFileHandler cannot be instantiated."""
        with self.assertRaises(TypeError):
            AbstractFileHandler("dummy.txt")

if __name__ == '__main__':
    unittest.main()
